+++ src/Game/PLUG/gameDrawer/GameRenderOffsetPass.java
@@ -2,5 +2,5 @@ package Game.PLUG.gameDrawer;
 
 @FunctionalInterface
 public interface GameRenderOffsetPass {
-    void passOffset(int offset);
+    void passOffset(float offset);
 }
